requirements.txt
argparse~=1.4.0
beautifulsoup4~=4.11.1
bs4~=0.0.1
requests~=2.27.1
urllib3~=1.26.9
qrcode~=7.3.1
pillow~=9.1.0
imgkit~=1.2.2
selenium~=4.1.3